AIAdvisor/skills/solr-opensearch-migration-advisor/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "solr-to-opensearch"
version = "0.2.0"
description = "Agent skill for migrating from Apache Solr to OpenSearch"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
keywords = ["opensearch", "solr", "migration", "agent", "skill"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "flake8>=7.0",
    "pyyaml>=6.0",
    "mcp",
]

[tool.setuptools.packages.find]
where = ["scripts"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["scripts", "tests/scenario-evals"]
